+++ src/app/api/tags/index.ts
@@ -1,0 +1,6 @@
+import { deleteTag } from "./deletetag";
+import { getTags } from "./getTags";
+import { postTag } from "./postTag";
+import { putTag } from "./putTag";
+
+export { getTags, postTag, putTag, deleteTag };
