+++ src/app/api/users/index.ts
@@ -1,0 +1,3 @@
+import { getUser } from "./getUser";
+
+export { getUser };
